requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
boto3
python-dotenv
requests
streamlit
